chore(worker): extract shared test helpers (BASE, fetchWorker, login, authed) (#247)
Closes #247.

`worker/routes/auth.test.ts` and `worker/routes/notebooks.test.ts` each
re-declared the same plumbing at the top of the file: `BASE`,
`fetchWorker`, `login(email)`, `authed(...)`. Moves all four into
`worker/test-helpers.ts` alongside the existing `resetDb` /
`extractSessionCookie` / `testEnv`. Tests shrink by 25–50 lines each
without changing any assertions.

- `TEST_BASE` exported for the origin used by every fetchWorker call.
- `fetchWorker(req)` wraps `worker.fetch(req, testEnv, ctx)`.
- `login(email)` does request-link → peek D1 for the token → callback
  → extract cookie, throws on a broken fixture (signals a real auth
  regression if tests were previously masking it via `!` on the
  cookie-null path).
- `authed(path, cookie, init?)` builds a Request with the session
  cookie set.

No behaviour change; 75 worker tests pass unchanged.

Full canonical gate green:
- typecheck (SPA + Worker)
- lint, format:check
- 1205 SPA tests, 75 Worker tests
- build
